+++ src/actions/login.ts
@@ -1,14 +1,44 @@
 "use server";
 
+import { DEFAULT_LOGIN_REDIRECT } from "@/routes";
 import { LoginSchema } from "@/schemas";
+import { signIn } from "@/services/auth";
+import { AuthError } from "next-auth";
+import { isRedirectError } from "next/dist/client/components/redirect";
 import * as z from "zod";
 
 export const login = async (values: z.infer<typeof LoginSchema>) => {
-  const validateFields = LoginSchema.safeParse(values);
+  const validatedFields = LoginSchema.safeParse(values);
 
-  if (!validateFields.success) {
+  if (!validatedFields.success) {
     return { error: "Dados informados inválidos!" };
   }
 
-  return { success: "Email enviado!" };
+  const { email, password } = validatedFields.data;
+
+  try {
+    await signIn("credentials", {
+      email,
+      password,
+      redirectTo: DEFAULT_LOGIN_REDIRECT,
+    });
+
+    return { success: "Autenticado com sucesso!" };
+
+  } catch (error) {
+    if (error instanceof AuthError) {
+      switch (error.type) {
+        case "CredentialsSignin":
+          return { error: "Usuário ou senha inválidos!" };
+        default:
+          return { error: `Ocorreu um erro AuthError: ${error.message}` };
+      }
+    }
+
+    if (isRedirectError(error)) {
+      console.log(error)
+    }
+
+    throw error;
+  }
 };
