+++ src/middleware.ts
@@ -1,19 +1,42 @@
-import { NextRequest, NextResponse } from "next/server";
-import { getUrl } from "./lib/get-url";
+import NextAuth from "next-auth";
 
-export function middleware(request: NextRequest) {
-  const token = request.cookies.get('authjs.session-token');
-  const pathname = request.nextUrl.pathname;
+import {
+  DEFAULT_LOGIN_REDIRECT,
+  apiAuthPrefix,
+  authRoutes,
+  publicRoutes,
+} from "@/routes";
+import authConfig from "@/services/auth/auth-config";
 
-  if (pathname === '/auth' && token) {
-    return NextResponse.redirect(new URL(getUrl('/admin')));
+const { auth } = NextAuth(authConfig);
+
+export default auth((req) => {
+  const { nextUrl } = req;
+  const isLoggedIn = !!req.auth;
+
+  const isApiAuthRoute = nextUrl.pathname.startsWith(apiAuthPrefix);
+  const isPublicRoute = publicRoutes.includes(nextUrl.pathname);
+  const isAuthRoute = authRoutes.includes(nextUrl.pathname);
+
+  if (isApiAuthRoute) {
+    return null;
   }
 
-  if (pathname === '/admin' && !token) {
-    return NextResponse.redirect(new URL(getUrl('/auth')));
+  if (isAuthRoute) {
+    if (isLoggedIn) {
+      return Response.redirect(new URL(DEFAULT_LOGIN_REDIRECT, nextUrl));
+    }
+    return null;
   }
-}
 
+  if (!isLoggedIn && !isPublicRoute) {
+    return Response.redirect(new URL("/auth/login", nextUrl));
+  }
+
+  return null;
+});
+
+// Optionally, don't invoke Middleware on some paths
 export const config = {
-  matcher: ['/((?!api|_next/static|_next/img|favicon.ico).*)'],
-}
+  matcher: ["/((?!.+\\.[\\w]+$|_next).*)", "/", "/(api|trpc)(.*)"],
+};
