+++ src/services/auth/index.ts
@@ -1,25 +1,20 @@
 import NextAuth from "next-auth";
-import EmailProvider from "next-auth/providers/email";
 import { FirestoreAdapter } from "@auth/firebase-adapter";
 import { cert } from "firebase-admin/app";
+import authConfig from "@/services/auth/auth-config";
 
 export const {
   handlers: { GET, POST },
   auth,
+  signIn,
+  signOut,
 } = NextAuth({
-  pages: {
-    signIn: "/admin",
-    signOut: "/auth",
-    error: "/auth",
-    verifyRequest: "/auth",
-    newUser: "/admin",
+  callbacks: {
+    async jwt({ token }) {
+      console.log({ token });
+      return token;
+    },
   },
-  providers: [
-    EmailProvider({
-      server: process.env.EMAIL_SERVER,
-      from: process.env.EMAIL_FROM,
-    }),
-  ],
   adapter: FirestoreAdapter({
     credential: cert({
       projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
@@ -27,4 +22,7 @@ export const {
       privateKey: process.env.FIREBASE_PRIVATE_KEY!.replace(/\\n/g, "\n"),
     }),
   }),
+  ...authConfig,
 });
+
+// PAREI COM ERRO AQUI https://youtu.be/1MTyCvS05V4?t=10047
